--- Packages/SharedUI/Sources/SharedUI/TextColor.swift
@@ -1,10 +1,0 @@
-import SwiftUI
-
-public extension Color {
-    static var textColor: Color {
-        if UITraitCollection.current.userInterfaceStyle == .dark {
-            return .white
-        }
-        return .black
-    }
-}
